# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/hamburguesas": {
            "get": {
                "description": "Obtiene todas las hamburguesas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hamburguesas"
                ],
                "summary": "Obtener todas las hamburguesas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.Hamburguesa"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Crea una nueva hamburguesa con los datos proporcionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hamburguesas"
                ],
                "summary": "Crear nueva hamburguesa",
                "parameters": [
                    {
                        "description": "Datos de la hamburguesa",
                        "name": "hamburguesa",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Hamburguesa"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Hamburguesa"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden: Access denied",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/hamburguesas/id/{id}": {
            "get": {
                "description": "Obtiene una hamburguesa basada en el ID proporcionado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hamburguesas"
                ],
                "summary": "Obtener hamburguesa por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la hamburguesa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Hamburguesa"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "ID invalido",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Hamburguesa no encontrada",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/hamburguesas/nombre/{name}": {
            "get": {
                "description": "Obtiene una hamburguesa basada en el nombre proporcionado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hamburguesas"
                ],
                "summary": "Obtener hamburguesa por nombre",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre de la hamburguesa",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Hamburguesa"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Nombre inválido",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Hamburguesa no encontrada",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/hamburguesas/price/{price}": {
            "get": {
                "description": "Obtiene una lista de hamburguesas cuyo precio es mayor al proporcionado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hamburguesas"
                ],
                "summary": "Obtener hamburguesas por precio",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Precio mínimo de la hamburguesa",
                        "name": "price",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.Hamburguesa"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Precio inválido",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/hamburguesas/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Actualiza una hamburguesa existente basada en el ID proporcionado y los datos JSON enviados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hamburguesas"
                ],
                "summary": "Actualizar hamburguesa por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la hamburguesa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la hamburguesa a actualizar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Hamburguesa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "JSON inválido",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden: Access denied",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Hamburguesa no encontrada",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Elimina una hamburguesa basada en el ID proporcionado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hamburguesas"
                ],
                "summary": "Eliminar hamburguesa por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la hamburguesa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden: Access denied",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Hamburguesa no encontrada",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/entity.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.ErrorResponse": {
            "description": "Estructura de respuesta para los errores",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "entity.Hamburguesa": {
            "type": "object",
            "required": [
                "descripcion",
                "imgUrl",
                "ingredientes",
                "nombre",
                "price"
            ],
            "properties": {
                "descripcion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imgUrl": {
                    "type": "string"
                },
                "ingredientes": {
                    "$ref": "#/definitions/entity.Ingrediente"
                },
                "nombre": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "entity.Ingrediente": {
            "type": "object",
            "properties": {
                "bacon": {
                    "type": "integer"
                },
                "cebolla": {
                    "type": "integer"
                },
                "huevo": {
                    "type": "integer"
                },
                "lechuga": {
                    "type": "integer"
                },
                "pepino": {
                    "type": "integer"
                },
                "tomate": {
                    "type": "integer"
                }
            }
        },
        "entity.SuccessResponse": {
            "description": "Estructura de respuesta para solicitudes exitosas",
            "type": "object",
            "properties": {
                "data": {}
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}